requirements.txt
Django==3.2
quests==2.28.1
